refactoring: move shared code from AbsenceController to VacationOverviewService
