+++ src/main/java/org/synyx/urlaubsverwaltung/absence/api/AbsenceController.java
@@ -10,21 +10,12 @@ import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 import org.synyx.urlaubsverwaltung.api.ResponseWrapper;
 import org.synyx.urlaubsverwaltung.api.RestApiDateFormat;
-import org.synyx.urlaubsverwaltung.application.domain.Application;
-import org.synyx.urlaubsverwaltung.application.domain.ApplicationStatus;
-import org.synyx.urlaubsverwaltung.application.service.ApplicationService;
 import org.synyx.urlaubsverwaltung.person.Person;
 import org.synyx.urlaubsverwaltung.person.PersonService;
-import org.synyx.urlaubsverwaltung.sicknote.SickNote;
-import org.synyx.urlaubsverwaltung.sicknote.SickNoteService;
-import org.synyx.urlaubsverwaltung.util.DateUtil;
+import org.synyx.urlaubsverwaltung.statistics.vacationoverview.api.VacationOverviewService;
 
-import java.time.DateTimeException;
-import java.time.LocalDate;
-import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
-import java.util.stream.Collectors;
 
 
 @Api("Absences: Get all absences for a certain period")
@@ -33,16 +24,13 @@ import java.util.stream.Collectors;
 public class AbsenceController {
 
     private final PersonService personService;
-    private final ApplicationService applicationService;
-    private final SickNoteService sickNoteService;
+    private final VacationOverviewService vacationOverviewService;
 
     @Autowired
-    public AbsenceController(PersonService personService, ApplicationService applicationService,
-                             SickNoteService sickNoteService) {
+    public AbsenceController(PersonService personService, VacationOverviewService vacationOverviewService) {
 
         this.personService = personService;
-        this.applicationService = applicationService;
-        this.sickNoteService = sickNoteService;
+        this.vacationOverviewService = vacationOverviewService;
     }
 
     @ApiOperation(
@@ -71,105 +59,11 @@ public class AbsenceController {
             throw new IllegalArgumentException("No person found for ID=" + personId);
         }
 
-        List<DayAbsence> absences = new ArrayList<>();
-        Person person = optionalPerson.get();
+        Optional<DayAbsence.Type> typeFilter =  type == null ? Optional.empty() : Optional.of(DayAbsence.Type.valueOf(type));
 
-        LocalDate startDate;
-        LocalDate endDate;
-        try {
-            startDate = getStartDate(year, Optional.ofNullable(month));
-            endDate = getEndDate(year, Optional.ofNullable(month));
-        } catch (DateTimeException exception) {
-            throw new IllegalArgumentException(exception.getMessage());
-        }
-
-        if (type == null || DayAbsence.Type.valueOf(type).equals(DayAbsence.Type.VACATION)) {
-            absences.addAll(getVacations(startDate, endDate, person));
-        }
-
-        if (type == null || DayAbsence.Type.valueOf(type).equals(DayAbsence.Type.SICK_NOTE)) {
-            absences.addAll(getSickNotes(startDate, endDate, person));
-        }
+        List<DayAbsence> absences = vacationOverviewService.personsVacations(optionalPerson.get(), year, Optional.ofNullable(month), typeFilter);
 
         return new ResponseWrapper<>(new DayAbsenceList(absences));
     }
 
-
-    private static LocalDate getStartDate(Integer year, Optional<Integer> optionalMonth) {
-
-        return optionalMonth.map(s -> DateUtil.getFirstDayOfMonth(year, s))
-            .orElseGet(() -> DateUtil.getFirstDayOfYear(year));
-
-    }
-
-
-    private static LocalDate getEndDate(Integer year, Optional<Integer> optionalMonth) {
-
-        return optionalMonth.map(s -> DateUtil.getLastDayOfMonth(year, s))
-            .orElseGet(() -> DateUtil.getLastDayOfYear(year));
-
-    }
-
-
-    private List<DayAbsence> getVacations(LocalDate start, LocalDate end, Person person) {
-
-        List<DayAbsence> absences = new ArrayList<>();
-
-        List<Application> applications = applicationService.getApplicationsForACertainPeriodAndPerson(start, end,
-                    person)
-                .stream()
-                .filter(application ->
-                            application.hasStatus(ApplicationStatus.WAITING)
-                            || application.hasStatus(ApplicationStatus.TEMPORARY_ALLOWED)
-                            || application.hasStatus(ApplicationStatus.ALLOWED))
-                .collect(Collectors.toList());
-
-        for (Application application : applications) {
-            LocalDate startDate = application.getStartDate();
-            LocalDate endDate = application.getEndDate();
-
-            LocalDate day = startDate;
-
-            while (!day.isAfter(endDate)) {
-                if (!day.isBefore(start) && !day.isAfter(end)) {
-                    absences.add(new DayAbsence(day, application.getDayLength().getDuration(), application.getDayLength().toString(), DayAbsence.Type.VACATION,
-                            application.getStatus().name(), application.getId()));
-                }
-
-                day = day.plusDays(1);
-            }
-        }
-
-        return absences;
-    }
-
-
-    private List<DayAbsence> getSickNotes(LocalDate start, LocalDate end, Person person) {
-
-        List<DayAbsence> absences = new ArrayList<>();
-
-        List<SickNote> sickNotes = sickNoteService.getByPersonAndPeriod(person, start, end)
-                .stream()
-                .filter(SickNote::isActive)
-                .collect(Collectors.toList());
-
-        for (SickNote sickNote : sickNotes) {
-            LocalDate startDate = sickNote.getStartDate();
-            LocalDate endDate = sickNote.getEndDate();
-
-            LocalDate day = startDate;
-
-            while (!day.isAfter(endDate)) {
-                if (!day.isBefore(start) && !day.isAfter(end)) {
-                    absences.add(new DayAbsence(day, sickNote.getDayLength().getDuration(),
-                        sickNote.getDayLength().toString(), DayAbsence.Type.SICK_NOTE, "ACTIVE",
-                        sickNote.getId()));
-                }
-
-                day = day.plusDays(1);
-            }
-        }
-
-        return absences;
-    }
 }
